Returning correct number of found documents now.

+++ com.openexchange.index/src/com/openexchange/index/IndexResult.java
@@ -65,7 +65,7 @@ public interface IndexResult<V> {
      *
      * @return The total number of results.
      */
-    int getNumFound();
+    long getNumFound();
 
     /**
      * Gets the requested results as specified by {@link QueryParameters}'s paging.
